requirements.txt
flask
flask-restful
flask-sqlalchemy
flask-migrate
flask-script
pymysql
flask-jwt-extended
requests
pytest-cov
werkzeug==0.16.0